baltasar_training_v1/requirements.txt
pandas>=2.2
numpy>=1.26
scikit-learn>=1.5
matplotlib>=3.8
plotly>=5.24
streamlit>=1.42
PyYAML>=6.0
joblib>=1.4
jupyter>=1.0
notebook>=7.2
